[jit] split test_type_sharing
